readyset-data: Fix MySQL REAL data type conversion
MySQL REAL datatype is stored as a 8-bytes floating point number [0].
Currently when lowering we are converting it to Float, which is a
4-bytes floating point number. This commit changes the conversion to
Double, which is a 8-bytes floating point number.

Fixed DDL Vertical to use Binary Protocol when querying the tables
as the text protocol always bring return as Bytes and they get compared
as Strings instead of Double.

[0]: https://dev.mysql.com/doc/refman/8.4/en/storage-requirements.html#data-types-storage-reqs-numeric

Ref: 4444
Fixes: #1270
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/7614
Tested-by: Buildkite CI

+++ readyset-data/src/dialect.rs
@@ -97,7 +97,7 @@ impl Dialect {
         // TODO: Handle `real_as_float` mode.
         match self.engine {
             SqlEngine::PostgreSQL => DfType::Float,
-            SqlEngine::MySQL => DfType::Float,
+            SqlEngine::MySQL => DfType::Double,
         }
     }
